Refactor HomeMatic / Fix issue with 0.104/dev
