+++ homeassistant/components/homematic/cover.py
@@ -6,9 +6,9 @@ from homeassistant.components.cover import (
     ATTR_TILT_POSITION,
     CoverDevice,
 )
-from homeassistant.const import STATE_UNKNOWN
 
-from . import ATTR_DISCOVER_DEVICES, HMDevice
+from .const import ATTR_DISCOVER_DEVICES
+from .entity import HMDevice
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -23,7 +23,7 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
         new_device = HMCover(conf)
         devices.append(new_device)
 
-    add_entities(devices)
+    add_entities(devices, True)
 
 
 class HMCover(HMDevice, CoverDevice):
@@ -68,9 +68,9 @@ class HMCover(HMDevice, CoverDevice):
     def _init_data_struct(self):
         """Generate a data dictionary (self._data) from metadata."""
         self._state = "LEVEL"
-        self._data.update({self._state: STATE_UNKNOWN})
+        self._data.update({self._state: None})
         if "LEVEL_2" in self._hmdevice.WRITENODE:
-            self._data.update({"LEVEL_2": STATE_UNKNOWN})
+            self._data.update({"LEVEL_2": None})
 
     @property
     def current_cover_tilt_position(self):
