+++ homeassistant/components/homematic/light.py
@@ -12,7 +12,8 @@ from homeassistant.components.light import (
     Light,
 )
 
-from . import ATTR_DISCOVER_DEVICES, HMDevice
+from .const import ATTR_DISCOVER_DEVICES
+from .entity import HMDevice
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -29,7 +30,7 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
         new_device = HMLight(conf)
         devices.append(new_device)
 
-    add_entities(devices)
+    add_entities(devices, True)
 
 
 class HMLight(HMDevice, Light):
